LayoutTests/imported/w3c/web-platform-tests/css/cssom/preferred-stylesheet-reversed-order-expected.txt

This text should be green

FAIL Preferred stylesheet where insertion order is tree order assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
